+++ NEWS.md
@@ -3,6 +3,7 @@
 * Fix shapefile export for non-linestring results starting with NULLs (#111)
 * Fix missing .prj in shapefile export (#110)
 * Improve recognition of non-standard field types names by db lookup (#112)
+* Upgrade node-pg dependency to 2.6.2
 
 1.5.4 - 2013-10-01
 ------------------