+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix missing .prj in shapefile export (#110)
 * Improve recognition of non-standard field types names by db lookup (#112)
 * Upgrade node-pg dependency to 2.6.2
+* Drop support for cluster
 
 1.5.4 - 2013-10-01
 ------------------